Version 0.01.006.0002 - Continue refactoring speed ramping (slowdown/speedup), including new class NonNegativeDouble for using type safety to squash bugs. Passes some unit tests.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/aarre/src/DriveMotor.java
@@ -19,7 +19,7 @@ public class DriveMotor extends MotorTorqueNADO implements ConcreteMotorInterfac
 	}
 
 	public double getTicksPerInch() {
-		return (getTicksPerRevolution() * DRIVE_GEAR_REDUCTION) / (WHEEL_DIAMETER_INCHES * Math
+		return (getTicksPerRevolution().doubleValue() * DRIVE_GEAR_REDUCTION) / (WHEEL_DIAMETER_INCHES * Math
 				.PI);
 	}
